Ft final ui ux enhancements 166847644

+++ sleeka-maria/app/Http/Controllers/SubcategoryController.php
@@ -5,39 +5,29 @@ namespace App\Http\Controllers;
 use App\Subcategory;
 use Illuminate\Http\Request;
 use App\Category;
+use Illuminate\Support\Facades\Auth;
 
 class SubcategoryController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index()
     {
-        $subcategories = Subcategory::latest()->get();
-        return $subcategories;
+        if(Auth::user()->role = 'Admin'){
+        $subcategories = Subcategory::latest()->paginate(10);
+        $categories = Category::all();
+        $count = 1;
+        return view('admin.subcategory',compact(['subcategories', 'categories', 'count']));
+        }
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+    
     public function store(Request $request)
-    {
-        
+    { 
+        if(Auth::user()->role = 'Admin'){  
         $this->validate($request,[
             'subcategory_name' => 'required',
             'category_id' => 'required|integer'
@@ -50,62 +40,37 @@ class SubcategoryController extends Controller
         $subcategory->subcategory_name = $request->subcategory_name;
         $subcategory->category_id = $category_id;
         $category->subcategories()->save($subcategory);
-        return $subcategory;
+        return back()->with('success','SubCategory Created');
+        }
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Subcategory  $subcategory
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Subcategory $subcategory)
+    
+    public function update(Request $request, $id)
     {
-        return $subcategory;
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Subcategory  $subcategory
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Subcategory $subcategory)
-    {
-        return $subcategory;
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Subcategory  $subcategory
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, Subcategory $subcategory)
-    {
+        if(Auth::user()->role = 'Admin'){
+        //dd($request->all());
         $this->validate($request, [
             'subcategory_name' => 'required',
-            'category_id' => 'required'|'integer'
+            'category_id' => 'required|integer'
         ]);
 
+        $subcategory = Subcategory::findOrFail($request->subcategory_id);
         $category_id = $request->category_id;
         $category = Category::findOrFail($category_id);
         $subcategory->subcategory_name = $request->subcategory_name;
         $subcategory->category_id = $category_id;
-        $category->subcategories()->update($subcategory);
-        return $subcategory;
+        $category->subcategories()->save($subcategory);
+        return redirect(route('subcategories.index'))->with('success','SubCategory Updated');
+        }
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Subcategory  $subcategory
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(Subcategory $subcategory)
+    
+    public function destroy(Request $request, $id)
     {
+        if(Auth::user()->role = 'Admin'){
+        $subcategory = Subcategory::findOrFail($request->category_id);
         $subcategory->delete();
-        return $subcategory;
+        return redirect(route('subcategories.index'))->with('success','SubCategory Removed');
+        }
     }
 }
